fixes transactions seeding, and merchants items

+++ db/migrate/20181004004149_create_transactions.rb
@@ -2,7 +2,7 @@ class CreateTransactions < ActiveRecord::Migration[5.1]
   def change
     create_table :transactions do |t|
       t.integer :invoice_id
-      t.integer :credit_card_number
+      t.bigint :credit_card_number
       t.string :result
 
       t.timestamps
